requirements.txt
requests>=2.28.0
python-dotenv>=0.19.0
supabase>=1.0.0
flask>=2.3.0
flask-cors>=6.0.0
gunicorn>=20.1.0
